+++ helper_classes.py
@@ -76,10 +76,10 @@ class TreeParams():
 		from Settings.constants import constants
 		from Settings.arguments import arguments
 		first_node = Node()
-		first_node.board = np.zeros([])
+		first_node.board = np.zeros([], dtype=int)
 		first_node.street = 1
 		first_node.current_player = constants.players.P1
-		first_node.bets = np.array([arguments.ante, arguments.ante])
+		first_node.bets = np.array([arguments.ante, arguments.ante], dtype=int)
 		self.root_node = first_node
 		self.limit_to_street = True
